화면 컴포넌트 구조 조정
