+++ frontend/src/page/space/register/RegisterPage2.jsx
