+++ frontend/src/page/space/register/RegisterPage6.jsx
