+++ frontend/src/page/space/register/RegisterStepper.jsx
